pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cube-registry-dev"
version = "0.1.0"
description = "Development dependencies for cube-registry CI scripts and tests"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
    "jsonschema>=4.0",
    "jinja2>=3.1",
    "responses>=0.25",   # HTTP mocking for health check tests
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
